refactor(store): split module
