--- src/store/constants.ts
@@ -1,23 +1,0 @@
-export const SET_LOCALE = "SET_LOCALE";
-export const CHANGE_THEME = "CHANGE_THEME";
-export const SET_SESSION_ID = "SET_SESSION_ID";
-export const SET_TORRENTS = "SET_TORRENTS";
-export const SET_SESSION = "SET_SESSION";
-export const SET_SESSION_STATS = "SET_SESSION_STATS";
-export const TOGGLE_MENUOPEN = "TOGGLE_MENUOPEN";
-export const SET_SELECTED_IDS = "SET_SELECTED_IDS";
-
-// Add Torrent
-export const TOGGLE_ADD_TORRENT_DIALOG = "TOGGLE_ADD_TORRENT_DIALOG";
-export const SHOW_TORRENT_DOWNLOAD_OPTIONS = "SHOW_TORRENT_DOWNLOAD_OPTIONS";
-export const CLOSE_TORRENT_DOWNLOAD_OPTIONS = "CLOSE_TORRENT_DOWNLOAD_OPTIONS";
-export const SET_DOWNLOAD_FILES = "SET_DOWNLOAD_FILES";
-export const SET_DOWNLOAD_SELECTED_FILES = "SET_DOWNLOAD_SELECTED_FILES";
-export const SET_FREE_DISK_SPACE = "SET_FREE_DISK_SPACE";
-
-// Remove Torrents
-export const TOGGLE_REMOVE_TORRENTS_DIALOG = "TOGGLE_REMOVE_TORRENTS_DIALOG";
-
-// Torrent Detail
-export const SHOW_TORRENT_DETAIL = "SHOW_TORRENT_DETAIL";
-export const HIDE_TORRENT_DETAIL = "HIDE_TORRENT_DETAIL";
