+++ src/store/constants/add.ts
@@ -1,0 +1,6 @@
+export const TOGGLE_ADD_TORRENT_DIALOG = "TOGGLE_ADD_TORRENT_DIALOG";
+export const SHOW_TORRENT_DOWNLOAD_OPTIONS = "SHOW_TORRENT_DOWNLOAD_OPTIONS";
+export const CLOSE_TORRENT_DOWNLOAD_OPTIONS = "CLOSE_TORRENT_DOWNLOAD_OPTIONS";
+export const SET_DOWNLOAD_FILES = "SET_DOWNLOAD_FILES";
+export const SET_DOWNLOAD_SELECTED_FILES = "SET_DOWNLOAD_SELECTED_FILES";
+export const SET_FREE_DISK_SPACE = "SET_FREE_DISK_SPACE";
