+++ src/store/constants/detail.ts
@@ -1,0 +1,2 @@
+export const SHOW_TORRENT_DETAIL = "SHOW_TORRENT_DETAIL";
+export const HIDE_TORRENT_DETAIL = "HIDE_TORRENT_DETAIL";
