+++ src/store/constants/menu.ts
@@ -1,0 +1,1 @@
+export const TOGGLE_MENUOPEN = "TOGGLE_MENUOPEN";
