+++ src/store/selector/menu.ts
@@ -1,0 +1,3 @@
+import { IState } from "src/types";
+
+export const getMenuOpen = (state: IState) => state.menu.open;
