+++ src/utils/formatter.test.ts
@@ -71,7 +71,7 @@ describe.each`
 
 test("formatSize", () => {
   store.getState = jest.fn().mockReturnValue({
-    rpc: {
+    session: {
       session: {
         units: {
           sizeBytes: 1000,
@@ -86,7 +86,7 @@ test("formatSize", () => {
 
 test("formatSize format zero", () => {
   store.getState = jest.fn().mockReturnValue({
-    rpc: {
+    session: {
       session: {
         units: {
           sizeBytes: 1000,
@@ -102,7 +102,7 @@ test("formatSize format zero", () => {
 
 test("formatSpeed", () => {
   store.getState = jest.fn().mockReturnValue({
-    rpc: {
+    session: {
       session: {
         units: {
           speedBytes: 1000,
@@ -117,7 +117,7 @@ test("formatSpeed", () => {
 
 test("formatSpeed format zero", () => {
   store.getState = jest.fn().mockReturnValue({
-    rpc: {
+    session: {
       session: {
         units: {
           speedBytes: 1000,
