+++ src/utils/formatter.ts
@@ -4,6 +4,8 @@ import { find } from "lodash";
 import store from "src/store";
 import { LANGUAGES, DEFAULT_LANGUAGE } from "src/constants";
 
+import { getSessionSelector } from "src/store/selector/session";
+
 export const formatBytes = (
   number: number,
   bytes: number,
@@ -20,7 +22,7 @@ export const formatBytes = (
 };
 
 export const formatSize = (number: number, allowZero = false): string => {
-  const { session } = store.getState().rpc;
+  const session = getSessionSelector(store.getState());
   const { sizeBytes, sizeUnits } = session.units;
   if (!allowZero && !number) {
     return " ";
@@ -29,7 +31,7 @@ export const formatSize = (number: number, allowZero = false): string => {
 };
 
 export const formatSpeed = (number: number, allowZero = false): string => {
-  const { session } = store.getState().rpc;
+  const session = getSessionSelector(store.getState());
   const { speedBytes, speedUnits } = session.units;
   if (!allowZero && !number) {
     return " ";
